Runtime Fix

+++ src/mem.cpp
@@ -4,7 +4,7 @@
 // allocate on 16 byte boundary
 void* ialloc(int size) {
 	char* ptr = (char*)malloc(size + 16 + 16);
-	char* data = (ptr + 16 + 16 - (int)ptr % 16);
+	char* data = (ptr + 16 + 16 - (long long)ptr % 16);
 	char* info = (char*)data - 16; // 16 bytes in size
 	*(char**)info = ptr;
 	return data;
